+++ flutterfast/lib/features/analytics/services/amplitude_analytics_service.dart
@@ -52,7 +52,7 @@ class AmplitudeAnalyticsService extends FastAnalyticsService {
   @override
   void updateVersionId(String? versionId) {
     if (kIsWeb) {
-      amplitude.setUserProperties({'version': versionId});
+      amplitude.setUserProperties({'app_version': versionId});
     }
   }
 }