+++ flutterfast/lib/features/analytics/services/fast_analytics_service.dart
@@ -5,6 +5,7 @@ abstract class FastAnalyticsService {
 
   void updateUserId(String? userId);
 
+  /// Update app version ID on the User Properties
   void updateVersionId(String? versionId);
 
   void logEvent(String eventName, {Map<String, dynamic>? eventProperties});