+++ lib/src/commands/create_app/steps/cleanup/clear_unused_analytics_files.dart
@@ -6,17 +6,35 @@ Future<void> clearUnusedAnalyticsFiles(String analytics) async {
   Directory analyticsDirectory = Directory('lib/features/analytics/services');
   File amplitudeAnalyticsServiceFile = File('${analyticsDirectory.path}/amplitude_analytics_service.dart');
   File posthogAnalyticsServiceFile = File('${analyticsDirectory.path}/posthog_analytics_service.dart');
+  File firebaseAnalyticsServiceFile = File('${analyticsDirectory.path}/firebase_analytics_service.dart');
 
   if (analytics == 'amplitude') {
     await posthogAnalyticsServiceFile.delete();
     removeFeatureFromFile('Posthog', 'lib/main.dart');
     removeFeatureFromFile('Posthog', 'web/index.html');
     removeFeatureFromFile('Posthog', 'lib/app/get_it.dart');
+    await firebaseAnalyticsServiceFile.delete();
+    removeFeatureFromFile('Fanalytics', 'lib/main.dart');
+    removeFeatureFromFile('Fanalytics', 'web/index.html');
+    removeFeatureFromFile('Fanalytics', 'lib/app/get_it.dart');
   } else if (analytics == 'posthog') {
     await amplitudeAnalyticsServiceFile.delete();
     removeFeatureFromFile('Amplitude', 'lib/main.dart');
     removeFeatureFromFile('Amplitude', 'web/index.html');
     removeFeatureFromFile('Amplitude', 'lib/app/get_it.dart');
+    await firebaseAnalyticsServiceFile.delete();
+    removeFeatureFromFile('Fanalytics', 'lib/main.dart');
+    removeFeatureFromFile('Fanalytics', 'web/index.html');
+    removeFeatureFromFile('Fanalytics', 'lib/app/get_it.dart');
+  } else if (analytics == 'firebase') {
+    await amplitudeAnalyticsServiceFile.delete();
+    removeFeatureFromFile('Amplitude', 'lib/main.dart');
+    removeFeatureFromFile('Amplitude', 'web/index.html');
+    removeFeatureFromFile('Amplitude', 'lib/app/get_it.dart');
+    await posthogAnalyticsServiceFile.delete();
+    removeFeatureFromFile('Posthog', 'lib/main.dart');
+    removeFeatureFromFile('Posthog', 'web/index.html');
+    removeFeatureFromFile('Posthog', 'lib/app/get_it.dart');
   } else {
     stdout.writeln('Unknown Analytics Plaform: $analytics');
   }