+++ lib/src/commands/create_app/steps/cleanup/update_analytics_environment.dart
@@ -1,0 +1,23 @@
+import 'dart:io';
+
+Future<void> updateAnalyticsEnvironment(String analytics) async {
+  File analyticsEnvironment = File('lib/app/get_it.dart');
+
+  Map<String, String> analyticsMap = {
+    'amplitude': "amplitude",
+    'posthog': "posthog)",
+    'firebase': "fanalytics",
+  };
+
+  String analyticsValue = analyticsMap[analytics] ?? "amplitude";
+
+  try {
+    var content = await analyticsEnvironment.readAsString();
+
+    var modifiedContent = content.replaceAll("defaultValue: 'amplitude'),", "defaultValue: '$analyticsValue'),");
+
+    await analyticsEnvironment.writeAsString(modifiedContent);
+  } catch (e) {
+    stderr.writeln(e);
+  }
+}